+++ CHANGES.md
@@ -1,1 +1,2 @@
 - Fix experimentation mode for sparks if you have some in your bags
+- Fix crafting window position when opening the auction house with TradeSkillMaster enabled